% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p53.R
\name{baby_function}
\alias{baby_function}
\title{Print a message to the terminal.}
\usage{
baby_function()
}
\arguments{
\item{None}{Function doesn't take any arguments, yet.}
}
\value{
None Function prints to terminal but doesn't return anything.
}
\description{
More interesting description of what the function does.  Blah blah.
}
\examples{
baby_function()
}
\seealso{
The general \code{\link{print}} function.
}

